Removed old ui manager
