+++ src/app/components/index.ts
@@ -14,7 +14,6 @@ export * from "./draggable";
 export * from "./selecting";
 export * from "./split_panes";
 export * from "./zoomable";
-export * from "./color_picker";
 export * from "./gradient_picker";
 export * from "./minimizable_panel";
 export * from "./error_boundary";
