+++ src/app/utils/index.ts
@@ -124,8 +124,7 @@ export function readFileAsDataUrl(file: File): Promise<string> {
 }
 
 export function getExtensionFromFileName(filename: string) {
-  // eslint-disable-next-line
-  const m = filename.match(/\.([^\.]+)$/);
+  const m = filename.match(/\.([^.]+)$/);
   if (m) {
     return m[1].toLowerCase();
   } else {
@@ -134,8 +133,7 @@ export function getExtensionFromFileName(filename: string) {
 }
 
 export function getFileNameWithoutExtension(filename: string) {
-  // eslint-disable-next-line
-  return filename.replace(/\.([^\.]+)$/, "");
+  return filename.replace(/\.([^.]+)$/, "");
 }
 
 export function showOpenFileDialog(accept?: string[]): Promise<File> {
