+++ src/app/views/panels/widgets/fluent_mapping_editor.tsx
@@ -16,7 +16,7 @@ import { ColorPicker } from "../../../components/fluentui_color_picker";
 import { ContextedComponent } from "../../../context_component";
 import { isKindAcceptable, type2DerivedColumns } from "../../dataset/common";
 import { ScaleEditor } from "../scale_editor";
-import { CharticulatorPropertyAccessors, DropZoneView } from "./manager";
+import { CharticulatorPropertyAccessors, DropZoneView } from "./types";
 import { AppStore } from "../../../stores";
 import { ScaleValueSelector } from "../scale_value_selector";
 import { FunctionCall } from "../../../../core/expression";
