+++ src/app/views/panels/widgets/fluentui_manager.tsx
@@ -37,7 +37,7 @@ import {
 } from "../../../utils/index";
 import { DataFieldSelectorValue } from "../../dataset/data_field_selector";
 import { ReorderListView } from "../object_list_editor";
-import { InputColorGradient, FluentComboBoxFontFamily } from "./controls";
+import { FluentComboBoxFontFamily } from "./controls";
 import { GroupByEditor } from "./groupby_editor";
 import {
   ChartTemplate,
@@ -69,7 +69,7 @@ import {
   TooltipHost,
 } from "@fluentui/react";
 import { FluentMappingEditor } from "./fluent_mapping_editor";
-import { CharticulatorPropertyAccessors } from "./manager";
+import { CharticulatorPropertyAccessors } from "./types";
 import { FluentInputColor } from "./controls/fluentui_input_color";
 import { FluentInputExpression } from "./controls/fluentui_input_expression";
 
@@ -116,6 +116,8 @@ import { FluentUIGradientPicker } from "../../../components/fluent_ui_gradient_p
 import { OrderMode } from "../../../../core/specification/types";
 import { CustomCollapsiblePanel } from "./controls/custom_collapsible_panel";
 import { FluentUIReorderStringsValue } from "./controls/fluentui_reorder_string_value";
+import { InputColorGradient } from "./controls/input_gradient";
+import { dropdownStyles, onRenderOption, onRenderTitle } from "./styles";
 
 export type OnEditMappingHandler = (
   attribute: string,
@@ -513,46 +515,6 @@ export class FluentUIWidgetManager
   ) {
     const theme = getTheme();
     if (options.type == "dropdown") {
-      const iconStyles: CSSProperties = { marginRight: "8px" };
-
-      const onRenderOption = (option: IDropdownOption): JSX.Element => {
-        return (
-          <>
-            {option.data && option.data.icon && (
-              <FluentDropdown>
-                <Icon
-                  style={iconStyles}
-                  iconName={option.data.icon}
-                  aria-hidden="true"
-                  title={option.data.icon}
-                />
-              </FluentDropdown>
-            )}
-            <span>{option.text}</span>
-          </>
-        );
-      };
-
-      const onRenderTitle = (options: IDropdownOption[]): JSX.Element => {
-        const option = options[0];
-
-        return (
-          <FluentDropdownWrapper>
-            {option.data && option.data.icon && (
-              <FluentDropdown>
-                <Icon
-                  style={iconStyles}
-                  iconName={option.data.icon}
-                  aria-hidden="true"
-                  title={option.data.icon}
-                />
-              </FluentDropdown>
-            )}
-            <span>{option.text}</span>
-          </FluentDropdownWrapper>
-        );
-      };
-
       return (
         <Dropdown
           key={`${this.getKeyFromProperty(property)}-${options.label}-${
@@ -586,16 +548,7 @@ export class FluentUIWidgetManager
           }}
           styles={{
             ...defaultStyle,
-            title: {
-              ...defultComponentsHeight,
-              borderWidth: options.hideBorder ? "0px" : null,
-            },
-            dropdownItemsWrapper: {
-              minWidth: 90,
-            },
-            callout: {
-              marginTop: options.shiftCallout ? options.shiftCallout : null,
-            },
+            ...dropdownStyles(options),
           }}
         />
       );
@@ -1297,7 +1250,6 @@ export class FluentUIWidgetManager
   }
 
   // Layout elements
-  // eslint-disable-next-line max-lines-per-function
   public sectionHeader(
     title: string,
     widget?: JSX.Element,
@@ -1623,11 +1575,7 @@ export class FluentUIWidgetManager
     );
   }
 
-  public table(
-    rows: JSX.Element[][],
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    options: Prototypes.Controls.TableOptions
-  ): JSX.Element {
+  public table(rows: JSX.Element[][]): JSX.Element {
     return (
       <table className="charticulator__widget-table">
         <tbody>
