+++ src/tests/test_app/color_picker.tsx
@@ -1,31 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 import * as React from "react";
-import { ColorPicker, GradientPicker } from "../../app/components";
-
-export class ColorPickerTestView extends React.Component<{}, {}> {
-  public render() {
-    return (
-      <div>
-        <div
-          style={{
-            background: "#eee",
-            border: "10px solid #aaa",
-            display: "inline-block",
-          }}
-        >
-          <ColorPicker
-            defaultValue={{ r: 117, g: 232, b: 75 }}
-            allowNull={true}
-            onPick={(value) => {
-              console.log(value);
-            }}
-          />
-        </div>
-      </div>
-    );
-  }
-}
+import { GradientPicker } from "../../app/components";
 
 export class GradientPickerTestView extends React.Component<{}, {}> {
   public render() {
@@ -47,6 +23,5 @@ export class GradientPickerTestView extends React.Component<{}, {}> {
 }
 
 export function register(f: any) {
-  f("ColorPicker", ColorPickerTestView);
   f("GradientPicker", GradientPickerTestView);
 }
